requirements.txt
aiogram==3.17.0
python-dotenv==1.0.1
SQLAlchemy==2.0.38
python-barcode[images]
